# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/wallet": {
      "get": {
        "operationId": "WalletController_getBalance",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WalletBalanceDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Coin and gem balances with their cash/coin value and live rates",
        "tags": [
          "wallet"
        ]
      }
    },
    "/api/wallet/rates": {
      "get": {
        "operationId": "WalletController_getRates",
        "parameters": [
          {
            "name": "gem_amount",
            "required": false,
            "in": "query",
            "description": "Gems to price. Returns exactly what converting them would pay.",
            "schema": {
              "example": 12000,
              "type": "number"
            }
          },
          {
            "name": "coin_amount",
            "required": false,
            "in": "query",
            "description": "Coins to price. Returns exactly what converting them would pay.",
            "schema": {
              "example": 500,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WalletRatesDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Coin, gem, convert and withdrawal rate definitions, with an optional exact preview",
        "tags": [
          "wallet"
        ]
      }
    },
    "/api/wallet/transactions": {
      "get": {
        "operationId": "WalletController_listTransactions",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Rows per page, max 100.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WalletTransactionDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List wallet transactions, newest first",
        "tags": [
          "wallet"
        ]
      }
    },
    "/api/wallet/convert": {
      "post": {
        "operationId": "WalletController_convert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConvertCurrencyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WalletDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Convert coins to gems or gems to coins",
        "tags": [
          "wallet"
        ]
      }
    },
    "/api/wallet/withdrawals": {
      "post": {
        "operationId": "WalletController_createWithdrawal",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWithdrawalDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/WithdrawalDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Request a payout of coins to real money",
        "tags": [
          "wallet"
        ]
      },
      "get": {
        "operationId": "WalletController_listWithdrawals",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "1-based page number.",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Rows per page, max 100.",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "required": [
                        "data",
                        "total"
                      ],
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/WithdrawalDto"
                          }
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List my withdrawal requests",
        "tags": [
          "wallet"
        ]
      }
    },
...
  },
  "info": {
    "title": "Coinzu API",
    "description": "Coinzu backend API",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "WalletBalanceRatesDto": {
        "type": "object",
        "properties": {
          "coins_per_usd": {
            "type": "number",
            "example": 1000
          },
          "coins_per_gem": {
            "type": "number",
            "example": 0.025
          },
          "min_withdrawal_coins": {
            "type": "number",
            "example": 5000
          }
        },
        "required": [
          "coins_per_usd",
          "coins_per_gem",
          "min_withdrawal_coins"
        ]
      },
      "WalletBalanceDto": {
        "type": "object",
        "properties": {
          "coin_balance": {
            "type": "number",
            "example": 12500
          },
          "gem_balance": {
            "type": "number",
            "example": 12000
          },
          "coin_value_usd": {
            "type": "string",
            "example": "12.50"
          },
          "gem_value_coins": {
            "type": "number",
            "example": 300
          },
          "can_withdraw": {
            "type": "boolean"
          },
          "rates": {
            "$ref": "#/components/schemas/WalletBalanceRatesDto"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "coin_balance",
          "gem_balance",
          "coin_value_usd",
          "gem_value_coins",
          "can_withdraw",
          "rates",
          "updated_at"
        ]
      },
      "RateCoinDto": {
        "type": "object",
        "properties": {
          "coins_per_usd": {
            "type": "number",
            "example": 1000
          },
          "usd_per_coin": {
            "type": "string",
            "example": "0.001000"
          },
          "definition": {
            "type": "string",
            "example": "1,000 coins = $1.00"
          }
        },
        "required": [
          "coins_per_usd",
          "usd_per_coin",
          "definition"
        ]
      },
      "RateGemDto": {
        "type": "object",
        "properties": {
          "coins_per_gem": {
            "type": "number",
            "example": 0.025
          },
          "gems_per_coin": {
            "type": "number",
            "example": 40
          },
          "definition": {
            "type": "string",
            "example": "40 gems = 1 coin"
          }
        },
        "required": [
          "coins_per_gem",
          "gems_per_coin",
          "definition"
        ]
      },
      "GemToCoinPreviewDto": {
        "type": "object",
        "properties": {
          "from_gems": {
            "type": "number",
            "example": 12000
          },
          "to_coins": {
            "type": "number",
            "example": 300
          }
        },
        "required": [
          "from_gems",
          "to_coins"
        ]
      },
      "GemToCoinDto": {
        "type": "object",
        "properties": {
          "rate": {
            "type": "number",
            "example": 0.025
          },
          "formula": {
            "type": "string",
            "example": "floor(gems * coins_per_gem)"
          },
          "definition": {
            "type": "string"
          },
          "preview": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/GemToCoinPreviewDto"
              }
            ]
          }
        },
        "required": [
          "rate",
          "formula",
          "definition",
          "preview"
        ]
      },
      "CoinToGemPreviewDto": {
        "type": "object",
        "properties": {
          "from_coins": {
            "type": "number",
            "example": 300
          },
          "to_gems": {
            "type": "number",
            "example": 12000
          }
        },
        "required": [
          "from_coins",
          "to_gems"
        ]
      },
      "CoinToGemDto": {
        "type": "object",
        "properties": {
          "rate": {
            "type": "number",
            "example": 0.025
          },
          "formula": {
            "type": "string",
            "example": "floor(coins / coins_per_gem)"
          },
          "definition": {
            "type": "string"
          },
          "preview": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CoinToGemPreviewDto"
              }
            ]
          }
        },
        "required": [
          "rate",
          "formula",
          "definition",
          "preview"
        ]
      },
      "RateConvertDto": {
        "type": "object",
        "properties": {
          "gem_to_coin": {
            "$ref": "#/components/schemas/GemToCoinDto"
          },
          "coin_to_gem": {
            "$ref": "#/components/schemas/CoinToGemDto"
          }
        },
        "required": [
          "gem_to_coin",
          "coin_to_gem"
        ]
      },
      "RateWithdrawalDto": {
        "type": "object",
        "properties": {
          "min_withdrawal_coins": {
            "type": "number",
            "example": 5000
          },
          "min_withdrawal_usd": {
            "type": "string",
            "example": "5.00"
          },
          "requires_kyc": {
            "type": "boolean"
          }
        },
        "required": [
          "min_withdrawal_coins",
          "min_withdrawal_usd",
          "requires_kyc"
        ]
      },
      "WalletRatesDto": {
        "type": "object",
        "properties": {
          "coin": {
            "$ref": "#/components/schemas/RateCoinDto"
          },
          "gem": {
            "$ref": "#/components/schemas/RateGemDto"
          },
          "convert": {
            "$ref": "#/components/schemas/RateConvertDto"
          },
          "withdrawal": {
            "$ref": "#/components/schemas/RateWithdrawalDto"
          }
        },
        "required": [
          "coin",
          "gem",
          "convert",
          "withdrawal"
        ]
      },
      "WalletTransactionDto": {
        "type": "object",
        "properties": {
          "cz_wallet_transaction_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "currency": {
            "type": "string",
            "enum": [
              "coin",
              "gem"
            ]
          },
          "type": {
            "type": "string",
            "enum": [
              "earn",
              "spend",
              "withdrawal",
              "convert_in",
              "convert_out",
              "reversal"
            ]
          },
          "amount": {
            "type": "number",
            "example": 1000,
            "description": "Signed: positive credits, negative debits."
          },
          "balance_after": {
            "type": "number",
            "example": 13000
          },
          "source_type": {
            "type": "string",
            "enum": [
              "offer",
              "daily_checkin",
              "referral",
              "game",
              "streak",
              "withdrawal",
              "redeem",
              "lucky_draw",
              "achievement",
              "challenge",
              "convert",
              "admin_adjustment",
              "offerwall",
              "welcome_bonus"
            ]
          },
          "source_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "note": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_wallet_transaction_id",
          "user_id",
          "currency",
          "type",
          "amount",
          "balance_after",
          "source_type",
          "source_id",
          "note",
          "created_at"
        ]
      },
      "WalletDto": {
        "type": "object",
        "properties": {
          "cz_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "coin_balance": {
            "type": "number",
            "example": 12500
          },
          "gem_balance": {
            "type": "number",
            "example": 12000
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_user_id",
          "coin_balance",
          "gem_balance",
          "updated_at"
        ]
      },
      "ConvertCurrencyDto": {
        "type": "object",
        "properties": {
          "from_currency": {
            "type": "string",
            "example": "coin",
            "enum": [
              "coin",
              "gem"
            ]
          },
          "amount": {
            "type": "number",
            "example": 500,
            "description": "Amount of from_currency to spend."
          }
        },
        "required": [
          "from_currency",
          "amount"
        ]
      },
      "WithdrawalDto": {
        "type": "object",
        "properties": {
          "cz_withdrawal_request_id": {
            "type": "string",
            "format": "uuid"
          },
          "user_id": {
            "type": "string",
            "format": "uuid"
          },
          "amount_coins": {
            "type": "number",
            "example": 5000
          },
          "amount_usd": {
            "type": "string",
            "example": "5.00"
          },
          "method": {
            "type": "string",
            "enum": [
              "paypal",
              "bank",
              "crypto"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected",
              "paid"
            ]
          },
          "reviewed_by": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "reviewed_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "rejection_reason": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "cz_withdrawal_request_id",
          "user_id",
          "amount_coins",
          "amount_usd",
          "method",
          "status",
          "reviewed_by",
          "reviewed_at",
          "rejection_reason",
          "created_at"
        ]
      },
      "CreateWithdrawalDto": {
        "type": "object",
        "properties": {
          "amount_coins": {
            "type": "number",
            "example": 5000,
            "description": "Coins to withdraw."
          },
          "method": {
            "type": "string",
            "example": "paypal",
            "enum": [
              "paypal",
              "bank",
              "crypto"
            ]
          },
          "destination_details": {
            "type": "object",
            "example": {
              "paypal_email": "user@coinzu.app"
            },
            "description": "Payout account details. Encrypted before it is stored."
          }
        },
        "required": [
          "amount_coins",
          "method",
          "destination_details"
        ]
      },
...
    }
  }
}